feat: 写真が描画範囲に入ったときだけロードする
